Implemented monthly transactions and progressbar
Monthly transactions like monthly costs or taxes etc. can now be easily
added to buildings. Also, the sum of monthly transactions is displayed
in the gui.
A new GuiElement has been added: the GuiProgressbar! It can display a
textured progressbar and text. The color of the bar can be changed.